Sprites done
To use sprite you create it after GL is loaded and save it anywhere, to be able for him to work properly. When you need to render you just call Render()
Two types of sprites: Linear and Circle. Linear will move by one frame after each render-call. TextureLayout need to be created in order to create LinearSprite. CircleSprite will draw pure white circle. Accuracy of the circle is a constructor-parameter which defaults to 16. This value is equal to number of sides of the polygon used to approximate the circle.