Block rotation
To Rotate block you just need to set Rotation property to needed angle value. Notice, that you can't rotate only one block, you rotate a whole family of blocks - blocks that are instances of that class. But you can create constructor that takes rotation angle as an argument of constructor and sets it's Rotation property to that value. Then any block, that actually is that rotated instance will be rotated. Also, you probably need to overwrite Name property and register that instance in BlockRegistry to be able to use this block in world map file